+++ Assets/GAS/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.3] - 2024-06-13
+
+添加了带level形参的ApplyGE方法。
+
+### Changed
+
+- 添加了带level形参的ApplyGE方法。ga
+
+
 ## [1.1.2] - 2024-06-12
 
 修复了部分bug。编辑器界面部分优化